Moving AppContent into a separate file
