+++ src/applications/letters/containers/LettersApp.jsx
@@ -1,78 +1,21 @@
 import React from 'react';
-import * as Sentry from '@sentry/browser';
-import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
 
-import backendServices from 'platform/user/profile/constants/backendServices';
-import { RequiredLoginView } from 'platform/user/authorization/components/RequiredLoginView';
-import { externalServices } from 'platform/monitoring/DowntimeNotification';
-import DowntimeBanner from 'platform/monitoring/DowntimeNotification/components/Banner';
-import CallVBACenter from 'platform/static-data/CallVBACenter';
+import backendServices from '@department-of-veterans-affairs/platform-user/profile/backendServices';
+import { RequiredLoginView } from '@department-of-veterans-affairs/platform-user/RequiredLoginView';
+import { externalServices } from '@department-of-veterans-affairs/platform-monitoring/DowntimeNotification';
+import { DowntimeBanner } from '@department-of-veterans-affairs/platform-monitoring/exports';
+
+import AppContent from '../components/AppContent';
 import { isLoadingFeatures } from '../selectors';
 
-const UNREGISTERED_ERROR = 'vets_letters_user_unregistered';
-
-// This needs to be a React component for RequiredLoginView to pass down
-// the isDataAvailable prop, which is only passed on failure.
-export class AppContent extends React.Component {
-  constructor(props) {
-    super(props);
-    if (props.isDataAvailable === false) {
-      Sentry.captureException(new Error(UNREGISTERED_ERROR));
-      this.state = { errorLogged: true };
-    } else {
-      this.state = { errorLogged: false };
-    }
-  }
-
-  // eslint-disable-next-line camelcase
-  UNSAFE_componentWillReceiveProps(nextProps) {
-    // only log isDataAvailable error if one isn't already logged
-    if (nextProps.isDataAvailable === false && !this.state.errorLogged) {
-      Sentry.captureException(new Error(UNREGISTERED_ERROR));
-      this.setState({ errorLogged: true });
-    }
-  }
-
-  render() {
-    if (this.props.isDataAvailable === false) {
-      return (
-        <div className="usa-grid">
-          <h1>VA letters and documents</h1>
-          <va-alert status="error">
-            We weren’t able to find information about your VA letters. If you
-            think you should be able to access this information, please{' '}
-            <CallVBACenter />
-          </va-alert>
-          <p className="vads-u-margin-bottom--4" />
-        </div>
-      );
-    }
-
-    if (!this.props.featureFlagsLoading) {
-      return (
-        <div data-testid="appContentChildren" className="usa-grid">
-          {this.props.children}
-        </div>
-      );
-    }
-    return (
-      <div className="vads-u-margin-y--5">
-        <va-loading-indicator
-          data-testid="feature-flags-loading"
-          message="Loading your information..."
-        />
-      </div>
-    );
-  }
-}
-
-export function LettersApp({ user, children, featureFlagsLoading }) {
+export function LettersApp({ children, featureFlagsLoading, user }) {
   return (
     <RequiredLoginView
-      verify
       serviceRequired={backendServices.EVSS_CLAIMS}
       user={user}
+      verify
     >
       <AppContent featureFlagsLoading={featureFlagsLoading}>
         <DowntimeBanner
@@ -85,22 +28,15 @@ export function LettersApp({ user, children, featureFlagsLoading }) {
   );
 }
 
-AppContent.propTypes = {
-  children: PropTypes.element,
-  isDataAvailable: PropTypes.bool,
-};
-
 LettersApp.propTypes = {
-  children: PropTypes.element,
+  children: PropTypes.node,
   featureFlagsLoading: PropTypes.bool,
-  user: PropTypes.shape({}),
+  user: PropTypes.object,
 };
 
-function mapStateToProps(state) {
-  return {
-    featureFlagsLoading: isLoadingFeatures(state),
-    user: state.user,
-  };
-}
+const mapStateToProps = state => ({
+  featureFlagsLoading: isLoadingFeatures(state),
+  user: state.user,
+});
 
 export default connect(mapStateToProps)(LettersApp);
